requirements.txt
asgiref==3.8.1
dj-database-url==2.2.0
Django==5.1.2
django-cors-headers==4.5.0
django-heroku==0.3.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
et-xmlfile==1.1.0
gunicorn==23.0.0
numpy==2.1.2
openpyxl==3.1.5
packaging==24.1
pandas==2.2.3
psycopg2==2.9.10
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
six==1.16.0
sqlparse==0.5.1
sqlparser==0.0.9
typing_extensions==4.12.2
tzdata==2024.2
whitenoise==6.7.0
